Adapt framework to the new auth force option
