+++ api/api/models/agent_inserted_model.py
@@ -2,12 +2,13 @@
 
 from __future__ import absolute_import
 
+from api.models.agent_added_model import AgentForce
 from api.models.base_model_ import Body
 
 
 class AgentInsertedModel(Body):
 
-    def __init__(self, id=None, name=None, ip=None, agent_id=None, key=None, force_time=None):
+    def __init__(self, id=None, name=None, ip=None, agent_id=None, key=None, force=None):
         """AgentAddedModel body model
         :param id: Agent id.
         :type id: str
@@ -19,8 +20,8 @@ class AgentInsertedModel(Body):
         :type agent_id: str
         :param key: Key to use when communicating with the manager. The agent must have the same key on its `client.keys` file.
         :type key: str
-        :param force_time: Remove the old agent with the same IP if disconnected since <force_time> seconds.
-        :type force_time: int
+        :param force: Remove the old agent with the same IP if disconnected since <force_time> seconds.
+        :type force: AgentForce
         """
         self.swagger_types = {
             'id': str,
@@ -28,7 +29,7 @@ class AgentInsertedModel(Body):
             'ip': str,
             'agent_id': str,
             'key': str,
-            'force_time': int
+            'force': AgentForce
         }
 
         self.attribute_map = {
@@ -37,7 +38,7 @@ class AgentInsertedModel(Body):
             'ip': 'ip',
             'agent_id': 'id',
             'key': 'key',
-            'force_time': 'force_time'
+            'force': 'force'
         }
 
         self._id = id
@@ -45,7 +46,7 @@ class AgentInsertedModel(Body):
         self._ip = ip
         self._agent_id = agent_id
         self._key = key
-        self._force_time = force_time
+        self._force = force
 
     @property
     def id(self) -> str:
@@ -123,16 +124,16 @@ class AgentInsertedModel(Body):
         self._key = key
 
     @property
-    def force_time(self) -> int:
+    def force(self) -> AgentForce:
         """
         :return: Limit time to disconnect an agent with the same IP.
-        :rtype: int
+        :rtype: AgentForce
         """
-        return self._force_time
+        return self._force
 
-    @force_time.setter
-    def force_time(self, force_time):
+    @force.setter
+    def force(self, force):
         """Limit time to disconnect an agent with the same IP.
-        :param force_time: Agents limit disconnection time. 
+        :param force: Agents limit disconnection time.
         """
-        self._force_time = force_time
+        self._force = force
