+++ framework/wazuh/core/agent.py
@@ -326,7 +326,7 @@ class Agent:
               'node_name': 'node_name', 'lastKeepAlive': 'last_keepalive', 'internal_key': 'internal_key',
               'registerIP': 'register_ip'}
 
-    def __init__(self, id=None, name=None, ip=None, key=None, force=-1, use_only_authd=False):
+    def __init__(self, id=None, name=None, ip=None, key=None, force=None, use_only_authd=False):
         """Initialize an agent.
 
         :param: id: When the agent exists
@@ -636,7 +636,7 @@ class Agent:
         except Exception as e:
             raise WazuhInternalError(1748, extra_message=str(e))
 
-    def _add(self, name, ip, id=None, key=None, force=-1, use_only_authd=False):
+    def _add(self, name, ip, id=None, key=None, force=None, use_only_authd=False):
         """Add an agent to Wazuh.
         2 uses:
             - name and ip [force]: Add an agent like manage_agents (generate id and key).
@@ -699,7 +699,7 @@ class Agent:
         except Exception as e:
             raise WazuhInternalError(1725, extra_message=str(e))
 
-    def _add_authd(self, name, ip, id=None, key=None, force=-1):
+    def _add_authd(self, name, ip, id=None, key=None, force=None):
         """Add an agent to Wazuh using authd.
         2 uses:
             - name and ip [force]: Add an agent like manage_agents (generate id and key).
@@ -715,7 +715,7 @@ class Agent:
             ID of the new agent.
         key : str
             Key of the new agent.
-        force : int
+        force : dict
             Remove old agents with same IP if disconnected since <force> seconds.
 
         Raises
@@ -740,14 +740,15 @@ class Agent:
         if key and len(key) < 64:
             raise WazuhError(1709)
 
-        force = force if type(force) == int else int(force)
-
         msg = ""
         if name and ip:
+            msg = {"function": "add", "arguments": {"name": name, "ip": ip}}
+
+            if force is not None:
+                msg["arguments"]["force"] = force
+
             if id and key:
-                msg = {"function": "add", "arguments": {"name": name, "ip": ip, "id": id, "key": key, "force": force}}
-            else:
-                msg = {"function": "add", "arguments": {"name": name, "ip": ip, "force": force}}
+                msg["arguments"].update({"id": id, "key": key})
 
         try:
             authd_socket = WazuhSocketJSON(common.AUTHD_SOCKET)
@@ -866,6 +867,7 @@ class Agent:
 
                                 # If force is non-negative then we check to remove the agent using value of force as
                                 # the max age in seconds
+                                # TODO refactor this with the new force
                                 if force >= 0 and Agent.check_if_delete_agent(entry_id, force):
                                     self.delete_agent_files(entry_id, entry_name, entry_ip, backup=True)
                                     # We add a void entry
